refactor(boot): dedupe wallet factory test context
Extract the shared SwingSet wallet-factory test context into
tools/walletFactoryContext.js, consumed by fast-usdc-deploy and
portfolio-deploy (which keep per-package published-path typing via thin
typed wrappers). Adds @aglocal/boot as a workspace dep of
fast-usdc-deploy.
