+++ packages/portfolio-deploy/test/walletFactory.ts
@@ -1,75 +1,26 @@
-import { Fail } from '@endo/errors';
-import {
-  type AgoricNamesRemotes,
-  makeAgoricNamesRemotesFromFakeStorage,
-} from '@agoric/vats/tools/board-utils.js';
-import {
-  fetchCoreEvalRelease,
-  makeSwingsetTestKit,
-} from '@aglocal/boot/tools/supports.js';
-import { makeWalletFactoryDriver } from '@aglocal/boot/tools/drivers.js';
+import { makeWalletFactoryContext as makeWalletFactoryContextShared } from '@aglocal/boot/tools/walletFactoryContext.js';
+import type { WalletFactoryDriver } from '@aglocal/boot/tools/drivers.js';
+import type { SwingsetTestKit } from '@aglocal/boot/tools/supports.js';
+import type { AgoricNamesRemotes } from '@agoric/vats/tools/board-utils.js';
 import type { PortfolioPublishedPathTypes } from '@aglocal/portfolio-contract/src/type-guards.ts';
 
-// XXX same as fast-usdc-deploy/test/walletFactory.ts and
-// XXX same as packages/boot/test/bootstrapTests/walletFactory.ts
-export const makeWalletFactoryContext = async (
-  t,
-  configSpecifier = '@agoric/vm-config/decentral-main-vaults-config.json',
-  opts = {} as Record<string, any>,
-) => {
-  const swingsetTestKit =
-    await makeSwingsetTestKit<PortfolioPublishedPathTypes>(t.log, undefined, {
-      configSpecifier,
-      ...opts,
-    });
-
-  const { runUtils, storage } = swingsetTestKit;
-  console.timeLog('DefaultTestContext', 'swingsetTestKit');
-  const { EV } = runUtils;
-
-  // Wait for ATOM to make it into agoricNames
-  await EV.vat('bootstrap').consumeItem('vaultFactoryKit');
-  console.timeLog('DefaultTestContext', 'vaultFactoryKit');
-
-  // has to be late enough for agoricNames data to have been published
-  const agoricNamesRemotes: AgoricNamesRemotes =
-    makeAgoricNamesRemotesFromFakeStorage(swingsetTestKit.storage);
-  const refreshAgoricNamesRemotes = () => {
-    Object.assign(
-      agoricNamesRemotes,
-      makeAgoricNamesRemotesFromFakeStorage(swingsetTestKit.storage),
-    );
-  };
-  agoricNamesRemotes.brand.ATOM || Fail`ATOM missing from agoricNames`;
-  console.timeLog('DefaultTestContext', 'agoricNamesRemotes');
-
-  const evalReleasedProposal = async (release: string, name: string) => {
-    const materials = await fetchCoreEvalRelease({
-      repo: 'Agoric/agoric-sdk',
-      release,
-      name,
-    });
-    try {
-      await swingsetTestKit.evalProposal(materials);
-    } finally {
-      refreshAgoricNamesRemotes();
-    }
-  };
-
-  const walletFactoryDriver = await makeWalletFactoryDriver(
-    runUtils,
-    storage,
-    agoricNamesRemotes,
-  );
-  return {
-    ...swingsetTestKit,
-    swingsetTestKit,
-    agoricNamesRemotes,
-    evalReleasedProposal,
-    refreshAgoricNamesRemotes,
-    walletFactoryDriver,
-  };
+type PortfolioSwingsetTestKit = SwingsetTestKit<PortfolioPublishedPathTypes>;
+type PortfolioWalletFactoryContext = PortfolioSwingsetTestKit & {
+  swingsetTestKit: PortfolioSwingsetTestKit;
+  agoricNamesRemotes: AgoricNamesRemotes;
+  evalReleasedProposal: (release: string, name: string) => Promise<void>;
+  refreshAgoricNamesRemotes: () => void;
+  walletFactoryDriver: WalletFactoryDriver;
 };
+
+type MakeWalletFactoryContext = (
+  t: unknown,
+  configSpecifier?: string,
+  opts?: Record<string, unknown>,
+) => Promise<PortfolioWalletFactoryContext>;
+
+export const makeWalletFactoryContext =
+  makeWalletFactoryContextShared as MakeWalletFactoryContext;
 
 export type WalletFactoryTestContext = Awaited<
   ReturnType<typeof makeWalletFactoryContext>
